[Refactoring ⚙️] useGetPost  구현 및 대체
