+++ src/components/PostDetail/PostDetail.tsx
@@ -4,11 +4,7 @@ import Modal from "../Modal/Modal"
 import PostDetailViewer from "./components/PostDetailViewer/PostDetailViewer"
 import PostDetailInfo from "./components/PostDetailInfo/PostDetailInfo"
 import { useParams } from "react-router-dom"
-import { useQuery } from "@tanstack/react-query"
-import { API } from "@/apis/Api"
-import { JSONPost, Post } from "@/types"
-
-const POST_DETAIL_QUERY_KEY = "POST_DETAIL_QUERY_KEY"
+import useGetPost from "./hooks/useGetPost"
 
 interface PostDetailProps {
   isShow: boolean
@@ -17,26 +13,7 @@ interface PostDetailProps {
 
 const PostDetail = ({ onClose, isShow }: PostDetailProps) => {
   const { id } = useParams()
-
-  const { data: post } = useQuery({
-    queryKey: [POST_DETAIL_QUERY_KEY, id],
-    queryFn: async () => {
-      return await API.get(`/posts/${id}`).then((res) => res.data)
-    },
-    select: (fetchPost: JSONPost) => {
-      const { content, mediaUrl, thumbnail } = JSON.parse(fetchPost.title)
-
-      const detailPost: Post = {
-        ...fetchPost,
-        title: {
-          mediaUrl: mediaUrl,
-          thumbnail: thumbnail,
-          content: content,
-        },
-      }
-      return detailPost
-    },
-  })
+  const post = useGetPost({ postId: id })
 
   if (!id) {
     onClose()
